Refactor: move test logic from per field test to general visibility test for RenderSingleField
